backend/requirements.txt
# Core web framework
fastapi==0.110.0
uvicorn==0.27.1

# Vector database and embeddings
qdrant-client==1.7.0
sentence-transformers==2.2.2

# Document processing
PyMuPDF==1.23.8
langdetect==1.0.9

# Data processing
numpy>=1.21.0
pandas>=1.3.0

# Configuration and utilities
python-dotenv==1.0.0
loguru==0.7.2
requests==2.31.0
pydantic==2.6.3
humps==0.2.2
pydantic-settings==2.2.1
huggingface-hub==0.19.4

# Self-hosted LLM dependencies
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
auto-gptq>=0.5.0

# Ollama integration
httpx>=0.25.0
